+++ scio-parquet/src/main/scala/com/spotify/scio/parquet/avro/package.scala
@@ -18,12 +18,6 @@
 package com.spotify.scio.parquet
 
 import com.spotify.scio.parquet.avro.syntax.Syntax
-import com.spotify.scio.util.ScioUtil
-import org.apache.avro.specific.SpecificRecord
-import org.apache.hadoop.conf.Configuration
-import org.apache.parquet.avro.{AvroDataSupplier, AvroReadSupport, GenericDataSupplier}
-
-import scala.reflect.ClassTag
 
 /**
  * Main package for Parquet Avro APIs. Import all.
@@ -39,31 +33,4 @@ package object avro extends Syntax {
 
   /** Alias for `me.lyh.parquet.avro.Predicate`. */
   val Predicate = me.lyh.parquet.avro.Predicate
-
-  def classIsSpecificRecord[T: ClassTag]: Boolean = {
-    val avroClass: Class[T] = ScioUtil.classOf[T]
-    classOf[SpecificRecord] isAssignableFrom avroClass
-  }
-
-  implicit class ConfigurationSyntax(conf: Configuration) {
-
-    def setReadSchemas[A, T](params: ParquetAvroIO.ReadParam[A, T]): Unit = {
-      AvroReadSupport.setAvroReadSchema(conf, params.readSchema)
-      if (params.projection != null) {
-        AvroReadSupport.setRequestedProjection(conf, params.projection)
-      }
-    }
-
-    def setDefaultGenericDataSupplier(): Unit = {
-      if (conf.get(AvroReadSupport.AVRO_DATA_SUPPLIER) == null) {
-        conf.setClass(
-          AvroReadSupport.AVRO_DATA_SUPPLIER,
-          classOf[GenericDataSupplier],
-          classOf[AvroDataSupplier]
-        )
-      }
-    }
-
-  }
-
 }
